sheets_requirements.txt
pandas>=1.5.0
gspread>=5.10.0
google-auth>=2.22.0
gspread-dataframe>=3.3.1 